+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [2.0.0] - 2023-08-14
 ### Added
 - behaviorDescription to components Addons
 - New title and paragraph elements to email editor
@@ -23,6 +23,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 - package-lock.json
 - webpack and vue cli deps
+
+## [1.7.0] - 2023-06-02
+### Added
+- Block level loop and conditional to advanced elements
+
+## [1.6.0] - 2023-06-01
+### Added
+- `canEdit` title option to settings
 
 ## [1.5.0] - 2023-02-23
 ### Added